Refactor syntax into implicit classes.
