+++ src/main/scala/com/mwz/sonar/scala/util/syntax/Config.scala
@@ -24,40 +24,35 @@ import java.nio.file.{Path, Paths}
 
 import cats.instances.string._
 import cats.syntax.eq._
+import com.mwz.sonar.scala.util.JavaOptionals._
 import org.sonar.api.config.Configuration
 
-import scala.language.implicitConversions
+object Config {
+  implicit final class ConfigOps(val configuration: Configuration) extends AnyVal {
 
-trait ConfigSyntax {
-  implicit final def configSyntax(configuration: Configuration): ConfigOps =
-    new ConfigOps(configuration)
-}
+    /**
+     * Get a list of paths for the given key.
+     * Fall back to the default value.
+     */
+    def getPaths(key: String, default: String): List[Path] =
+      configuration
+        .get(key)
+        .toOption
+        .filter(_.nonEmpty)
+        .getOrElse(default)
+        .split(',') // scalastyle:ignore org.scalastyle.scalariform.NamedArgumentChecker
+        .map(p => Paths.get(p.trim))
+        .toList
 
-final class ConfigOps(val configuration: Configuration) extends AnyVal {
-  import JavaOptionals._ // scalastyle:ignore org.scalastyle.scalariform.ImportGroupingChecker
-
-  /**
-   * Get a list of paths for the given key.
-   * Fall back to the default value.
-   */
-  def getPaths(key: String, default: String): List[Path] =
-    configuration
-      .get(key)
-      .toOption
-      .filter(_.nonEmpty)
-      .getOrElse(default)
-      .split(',') // scalastyle:ignore org.scalastyle.scalariform.NamedArgumentChecker
-      .map(p => Paths.get(p.trim))
-      .toList
-
-  /**
-   * Get a boolean property for the given key.
-   * Defaults to false.
-   */
-  def getValue[T](key: String)(implicit ev: T =:= Boolean): Boolean = {
-    configuration
-      .get(key)
-      .toOption
-      .exists(_.toLowerCase === "true")
+    /**
+     * Get a boolean property for the given key.
+     * Defaults to false.
+     */
+    def getValue[T](key: String)(implicit ev: T =:= Boolean): Boolean = {
+      configuration
+        .get(key)
+        .toOption
+        .exists(_.toLowerCase === "true")
+    }
   }
 }
